调整renderer代码结构
