+++ src/renderer/assets/variable.scss
@@ -8,3 +8,6 @@ $color-border2: #E4E7ED;
 $color-border3: #EBEEF5;
 $color-border4: #F2F6FC;
 $color-table-text: #606266;
+
+$searchBar-height: 45px;
+$leftmenu-width: 200px;
