+++ src/renderer/view/artist/index.vue
@@ -48,82 +48,7 @@
         </el-table>
     </div>
 </template>
-<script>
-    import {mapActions} from 'vuex'
-
-    export default {
-        data() {
-            return {
-                detail: {},
-                songs: [],
-                loading: false
-            }
-        },
-        computed: {
-            id() {
-                return this.$route.params.id
-            },
-            vendor() {
-                return this.$route.query.vendor
-            }
-        },
-        methods: {
-            ...mapActions('api', ['play']),
-            async getSongs() {
-                this.loading = true
-                try {
-                    let data = await Vue.$musicApi.getArtistSongs(this.vendor, this.id)
-                    if (data.status) {
-                        this.detail = data.data.detail
-                        this.songs = data.data.songs.map(item => {
-                            return {
-                                ...item,
-                                commentId: item.id,
-                                songId: item.id,
-                                vendor: this.vendor
-                            }
-                        })
-                    }
-                } catch (e) {
-                    console.warn(e)
-                    e.msg && this.$message.warning(e.msg)
-                }
-                this.loading = false
-            },
-            rowClassName({row, rowIndex}) {
-                const rs = [
-                    this.s.row
-                ]
-                if (row.cp) {
-                    rs.push(this.s.disabled)
-                }
-                return rs.join(' ')
-            },
-            doPlay(item) {
-                const list = []
-                this.songs.forEach(item => {
-                    list.push(item)
-                })
-                this.$store.commit('c_playlist/update', list)
-                console.log(item)
-                this.play(item)
-            },
-        },
-        created() {
-            this.getSongs()
-        },
-        beforeRouteUpdate(to, form, next) {
-            next()
-            this.getSongs()
-        },
-        beforeRouteEnter(to, from, next) {
-            if (!to.query.vendor) {
-                return Vue.router.push('/')
-            }
-            next()
-        }
-    }
-</script>
+<script src="./index.js"></script>
 <style lang="scss" module="s">
     .artist {
         .top {
