+++ src/renderer/view/comments/eventBus.js
@@ -1,9 +1,9 @@
 export default new Vue({
-  data() {
-    return {
-      hotComments: [],
-      comments: [],
-      total: 0
+    data() {
+        return {
+            hotComments: [],
+            comments: [],
+            total: 0
+        }
     }
-  }
 })
