+++ src/renderer/view/comments/titleBar.vue
@@ -2,14 +2,14 @@
     <div :class="s.titleBar">{{title}}</div>
 </template>
 <script>
-  export default {
-    props: {
-      title: {
-        type: String,
-        required: true
-      }
+    export default {
+        props: {
+            title: {
+                type: String,
+                required: true
+            }
+        }
     }
-  }
 </script>
 <style lang="scss" module="s">
     .titleBar {
