+++ src/renderer/view/components/current_playlist/index.vue
