+++ src/renderer/view/components/leftMenu/avatar.vue
