+++ src/renderer/view/components/leftMenu/offline-playlist/item.vue
