+++ src/renderer/view/components/leftMenu/playlist/index.vue
