+++ src/renderer/view/components/leftMenu/playlist/item.vue
