+++ src/renderer/view/components/leftMenu/viewControl.vue
