+++ src/renderer/view/components/lyrics/index.js
