+++ src/renderer/view/components/player/components/playControl/index.js
