+++ src/renderer/view/components/player/components/playControl/index.vue
