+++ src/renderer/view/components/player/components/playInfo/index.js
