+++ src/renderer/view/components/player/images/defaultAlbum.png
