+++ src/renderer/view/components/player/index.js
