+++ src/renderer/view/components/player/index.scss
