+++ src/renderer/view/components/player/index.vue
