+++ src/renderer/view/playlist/index.vue
